[feat-390] add _font and change imports
