+++ src/style/_fonts.ts
@@ -1,0 +1,10 @@
+export const fontFamilyString = [
+  '"Alef"',
+  '-apple-system',
+  'BlinkMacSystemFont',
+  '"Segoe UI"',
+  'Roboto',
+  '"Helvetica Neue"',
+  'Arial',
+  'sans-serif',
+].join(',');
