+++ src/style/_sizes.ts
@@ -7,16 +7,5 @@ export const cardWidth = 500;
 export const cardHeaderHeight = 70;
 export const cardFooterHeight = 50;
 
-export const fontFamilyString = [
-  '"Alef"',
-  '-apple-system',
-  'BlinkMacSystemFont',
-  '"Segoe UI"',
-  'Roboto',
-  '"Helvetica Neue"',
-  'Arial',
-  'sans-serif',
-].join(',');
-
 //road number icon && font
 export const roadIconFont = 'sans-serif';
