+++ src/style/index.ts
@@ -1,3 +1,4 @@
+export * from './_fonts';
 export * from './_colors';
 export * from './_sizes';
 export { defaultThemeOptions } from './theme';
